+++ src/fs/create.js
@@ -1,10 +1,10 @@
 import fs from 'fs/promises';
 import path from 'path';
-import { __dirname } from './custom-module.js';
+import { __dirname } from '../path.module.js';
 
 export const create = async () => {
   const data = 'I am fresh and young';
-  const sourcePath = path.resolve(__dirname, './files/fresh.txt');
+  const sourcePath = path.resolve(__dirname, './fs/files/fresh.txt');
 
   try {
     const stat = await fs.stat(sourcePath);
