--- src/fs/custom-module.js
@@ -1,9 +1,0 @@
-
-import path from 'path';
-
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-export { __filename, __dirname };
