+++ src/fs/delete.js
@@ -1,9 +1,9 @@
 import fs from 'fs/promises';
 import path from 'path';
-import { __dirname } from './custom-module.js';
+import { __dirname } from '../path.module.js';
 
 export const remove = async () => {
-  const removeFilePath = path.resolve(__dirname, './files/fileToRemove.txt');
+  const removeFilePath = path.resolve(__dirname, './fs/files/fileToRemove.txt');
 
   try {
     await fs.rm(removeFilePath);
