+++ src/fs/rename.js
@@ -1,10 +1,10 @@
 import fs from 'fs/promises';
 import path from 'path';
-import { __dirname } from './custom-module.js';
+import { __dirname } from '../path.module.js';
 
 export const rename = async () => {
-  const wrongFilePath = path.resolve(__dirname, './files/wrongFilename.txt');
-  const properFilePath = path.resolve(__dirname, './files/properFilename.md');
+  const wrongFilePath = path.resolve(__dirname, './fs/files/wrongFilename.txt');
+  const properFilePath = path.resolve(__dirname, './fs/files/properFilename.md');
 
   try {
     await fs.rename(wrongFilePath, properFilePath);
